+++ lib/ogurets.dart
@@ -26,10 +26,10 @@ part 'src/ogurets_opts.dart';
 part 'src/output/basic_formatter.dart';
 part 'src/output/console_buffer.dart';
 part 'src/output/delegating_formatter.dart';
+part 'src/output/filtered_formatter.dart';
 part 'src/output/formatter.dart';
 part 'src/output/intellij_formatter.dart';
 part 'src/output/output.dart';
 part "src/status/status.dart";
 part 'src/task.dart';
 
-final Logger _log = Logger('ogurets');