+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+4.0.5
+=====
+* Added option to filter out certain types of messages that are logged by default (steps, backgrounds, scenario's, features)
+
+4.0.4
+=====
+* Added option to change log level of ogurets through OguretsOpts
+
 4.0.3
 =====
 * Skip abstract classes containing step definitions